# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API de paiement mobile OM Pay avec authentification OTP",
        "contact": {
            "email": "support@ompay.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://annasock-ompay.onrender.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/comptes": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un compte utilisateur avec dépôt initial",
                "description": "Crée un nouvel utilisateur (ou récupère un existant via NCI), génère automatiquement un numéro de compte Orange Money, login et code marchand, puis effectue un dépôt initial",
                "operationId": "createCompte",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "nci",
                                    "email",
                                    "numero_user",
                                    "password",
                                    "montant_initial"
                                ],
                                "properties": {
                                    "nom": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Doe"
                                    },
                                    "prenom": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John"
                                    },
                                    "adresse": {
                                        "description": "Adresse de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "123 Rue de la Paix, Dakar"
                                    },
                                    "nci": {
                                        "description": "Numéro de carte d'identité nationale (unique)",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "1234567890123"
                                    },
                                    "email": {
                                        "description": "Adresse email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "john.doe@example.com"
                                    },
                                    "numero_user": {
                                        "description": "Numéro d'utilisateur unique",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "771234567"
                                    },
                                    "password": {
                                        "description": "Mot de passe (sera hashé automatiquement)",
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "password123"
                                    },
                                    "montant_initial": {
                                        "description": "Montant du dépôt initial",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données d'entrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "properties": {
                                                "nom": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Le nom est obligatoire."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Cette adresse email est déjà utilisée."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Détails de l'erreur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/{numero-compte}/solde": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Consulter le solde d'un compte",
                "description": "Récupère le solde actuel d'un compte Orange Money. L'utilisateur doit être authentifié et propriétaire du compte.",
                "operationId": "getSolde",
                "parameters": [
                    {
                        "name": "numero-compte",
                        "in": "path",
                        "description": "Numéro du compte bancaire",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1234567890"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "compte_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - L'utilisateur n'est pas propriétaire du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}